Update nUnit example
