+++ dotnet-nunit/SauceLabs.Demo.Visual.Example/SauceLabs.Demo.Visual.Example.csproj
@@ -2,7 +2,7 @@
 
   <PropertyGroup>
     <TargetFramework>net6.0</TargetFramework>
-    <RootNamespace>SauceLabs.Visual.Example</RootNamespace>
+    <RootNamespace>SauceLabs.Demo.Visual.Example</RootNamespace>
     <Nullable>enable</Nullable>
 
     <IsPackable>false</IsPackable>
@@ -13,7 +13,7 @@
     <PackageReference Include="NUnit" Version="3.13.2" />
     <PackageReference Include="NUnit3TestAdapter" Version="4.0.0" />
     <PackageReference Include="coverlet.collector" Version="3.1.0" />
-    <PackageReference Include="SauceLabs.Visual" Version="0.0.3" />
+    <PackageReference Include="SauceLabs.Visual" Version="0.1.0" />
     <PackageReference Include="Selenium.WebDriver" Version="4.10.0" />
     <PackageReference Include="WebDriverManager" Version="2.16.3" />
     <PackageReference Include="MSTest.TestAdapter" Version="2.2.10" />
