+++ dotnet-nunit/SauceLabs.Demo.Visual.Example/SetupFixture.cs
@@ -1,0 +1,15 @@
+using System.Threading.Tasks;
+using NUnit.Framework;
+using SauceLabs.Visual;
+
+namespace SauceLabs.Demo.Visual.Example;
+
+[SetUpFixture]
+public class SetupFixture
+{
+    [OneTimeTearDown]
+    public async Task RunAfterAnyTests()
+    {
+        await VisualClient.Finish();
+    }
+}
