+++ dotnet-nunit/nunit-example.sln
@@ -3,7 +3,7 @@ Microsoft Visual Studio Solution File, Format Version 12.00
 # Visual Studio Version 17
 VisualStudioVersion = 17.0.31903.59
 MinimumVisualStudioVersion = 10.0.40219.1
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "SauceLabs.Visual.Example", "SauceLabs.Visual.Example\SauceLabs.Visual.Example.csproj", "{23758973-F2DF-4E61-AC5B-591B6EAFE228}"
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "SauceLabs.Demo.Visual.Example", "SauceLabs.Demo.Visual.Example\SauceLabs.Demo.Visual.Example.csproj", "{23758973-F2DF-4E61-AC5B-591B6EAFE228}"
 EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
